+++ src/ui/pages/products/ProductsPage.h
@@ -27,7 +27,7 @@ private slots:
 private:
     void buildActions();
     void loadFromStorage();
-    unsigned short int computeNextId() const;
+    uint32_t computeNextId() const;
 
     ProductTableModel*     m_model;
     QSortFilterProxyModel* m_proxy;
